refactor: extract CLI helpers and rules command
